UPDATE : Refactor package structure
