+++ src/main/java/lotto/api/EmptyModel.java
@@ -1,14 +1,14 @@
-package lotto;
+package lotto.api;
 
 import java.util.HashMap;
 import java.util.Map;
 
-public class EmptyModel {
+class EmptyModel {
     private static Map<String, Object> emptyModel;
 
     private EmptyModel() {  }
 
-    public static Map<String, Object> get() {
+    static Map<String, Object> get() {
         if (emptyModel == null) {
             emptyModel = new HashMap<>();
         }
