+++ src/main/java/lotto/api/TemplateEngine.java
@@ -1,12 +1,12 @@
-package lotto;
+package lotto.api;
 
 import java.util.Map;
 
 import spark.ModelAndView;
 import spark.template.handlebars.HandlebarsTemplateEngine;
 
-public class TemplateEngine {
-    public static String render(Map<String, Object> model, String templatePath) {
+class TemplateEngine {
+    static String render(Map<String, Object> model, String templatePath) {
         return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
     }
 }
