+++ src/main/java/lotto/data/JDBCConnection.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.data;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
